@@ -1,6 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
+import { CatFoodService } from './cat-food.service';
 import { AppComponent } from './app.component';
 
 @NgModule({
@@ -11,7 +12,7 @@ import { AppComponent } from './app.component';
     BrowserModule, 
     HttpClientModule
   ],
-  providers: [],
+  providers: [CatFoodService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
